# 🛠️ Tool 6: OutputFormatter - Template-Based Output Generation