## Summary

Implements issue #113 - OutputFormatter tool for consistent, template-based output generation across all commands.

Replaces 200-300 lines of manual template specifications per command with 5-10 lines using Jinja2 templates.

## Features Implemented

### Core Functionality ✅
- **format_markdown()** - Markdown output with Jinja2 templates
- **format_json()** - JSON output with schema support
- **format_table()** - Markdown table generation
- **executive_summary()** - High-level executive summaries
- **technical_report()** - Detailed technical reports with code examples
- **stakeholder_update()** - Stakeholder-specific updates

### Template System ✅
- Base template with common macros and elements
- 7 production-ready templates:
  - `base.md.j2` - Shared macros and layout
  - `project_status.md.j2` - Project health and timeline reports
  - `decision_analysis.md.j2` - Decision analysis with options
  - `follow_up_report.md.j2` - Action item tracking
  - `health_dashboard.md.j2` - Health metrics dashboard
  - `stakeholder_update.md.j2` - Stakeholder communications
  - `executive_summary.md.j2` - Executive summaries
  - `technical_report.md.j2` - Technical reports

### Emoji & Styling ✅
- **apply_emoji_indicators()** - Health-based emoji application
- **format_date()** - Multiple date styles (relative, absolute, short, long, ISO)
- Health score formatting with emojis and percentages
- Consistent visual hierarchy and styling

### Performance ✅
- **<50ms** template rendering (after cache warmup)
- Template caching via Jinja2
- Performance tracking and statistics
- Meets all performance requirements

## Test Results

✅ **28 tests passing**
✅ **91% code coverage** (exceeds 80% requirement)
✅ **Performance verified**: <50ms rendering after warmup

```
============================== 28 passed in 1.59s ==============================
tools/output_formatter.py              155     14    91%
tools/output_models.py                  86      0   100%
```

## API Example

**Before (200+ lines of template specs):**
```python
# Manual template construction with 200+ lines...
output = f"# {title}\n\n"
output += f"## Health: {health_score}\n\n"
# ... 200 more lines of string formatting
```

**After (5-10 lines):**
```python
from tools import OutputFormatter

formatter = OutputFormatter()
output = formatter.format_markdown(
    health_data,
    template="project_status",
    context={"focus": "risks"}
)
print(output.content)
```

## File Structure

```
tools/
├── output_formatter.py      (155 lines, 91% coverage)
└── output_models.py          (86 lines, 100% coverage)

templates/output/
├── base.md.j2                (Common macros)
├── project_status.md.j2      (Project reports)
├── decision_analysis.md.j2   (Decision analysis)
├── follow_up_report.md.j2    (Action tracking)
├── health_dashboard.md.j2    (Health metrics)
├── stakeholder_update.md.j2  (Stakeholder comms)
├── executive_summary.md.j2   (Executive summaries)
└── technical_report.md.j2    (Technical reports)

tests/
└── test_output_formatter.py  (28 tests, all passing)
```

## Success Criteria Met

- ✅ All format types implemented (markdown, JSON, table)
- ✅ Test coverage >80% (achieved 91%)
- ✅ Documentation complete with docstrings
- ✅ All 7 templates created and tested
- ✅ Performance <50ms for template rendering
- ✅ Ready for integration into commands

## Next Steps

After merge, integrate OutputFormatter into:
1. `/project-status` command
2. `/follow-up-check` command
3. `/decide` command
4. All other commands using manual templates

## Benefits

- **200-300 lines → 5-10 lines** per command
- **Consistency**: Single source of truth for formatting
- **Maintainability**: Templates easier to update than code
- **Performance**: <50ms rendering with Jinja2 caching
- **Flexibility**: Support for multiple audiences and formats

---

🤖 Generated with [Claude Code](https://claude.com/claude-code)